refactor: migrate no-absolute-path rule
