+++ src/index.ts
@@ -30,6 +30,7 @@ import noDuplicates from './rules/no-duplicates'
 import first from './rules/first'
 import maxDependencies from './rules/max-dependencies'
 import noExtraneousDependencies from './rules/no-extraneous-dependencies'
+import noAbsolutePath from './rules/no-absolute-path'
 
 // configs
 import recommended from './config/recommended'
@@ -71,7 +72,7 @@ export const rules = {
   first,
   'max-dependencies': maxDependencies,
   'no-extraneous-dependencies': noExtraneousDependencies,
-  'no-absolute-path': require('./rules/no-absolute-path'),
+  'no-absolute-path': noAbsolutePath,
   'no-nodejs-modules': require('./rules/no-nodejs-modules'),
   'no-webpack-loader-syntax': require('./rules/no-webpack-loader-syntax'),
   order: require('./rules/order'),
